+++ anyblok/bloks/anyblok_core/system/blok.py
@@ -96,23 +96,6 @@ class Blok:
         """
         # Do not remove blok because 2 or More AnyBlok api may use the same
         # Database
-
-        Association = cls.registry.System.Blok.Association
-
-        def populate_association(blok, links, mode, createifnotexist=False):
-            for link in links:
-                if createifnotexist:
-                    if not cls.query().filter(cls.name == link).count():
-                        cls.insert(name=link, state='undefined',
-                                   version='None')
-                query = Association.query()
-                query = query.filter(Association.blok == blok)
-                query = query.filter(Association.linked_blok == link)
-                query = query.filter(Association.mode == mode)
-                if not query.count():
-                    Association.insert(blok=blok, linked_blok=link, mode=mode)
-
-        # Create blok if not exist
         for order, blok in enumerate(BlokManager.ordered_bloks):
             b = cls.query().filter(cls.name == blok)
             version = BlokManager.bloks[blok].version
@@ -121,16 +104,6 @@ class Blok:
                 b.version = version
             else:
                 cls.insert(name=blok, order=order, version=version)
-
-        # Update required, optional, conditional
-        for blok in BlokManager.ordered_bloks:
-            entry = BlokManager.bloks[blok]
-
-            populate_association(blok, entry.required, 'required')
-            populate_association(blok, entry.optional, 'optional',
-                                 createifnotexist=True)
-            populate_association(blok, entry.conditional, 'conditional',
-                                 createifnotexist=True)
 
     @classmethod
     def apply_state(cls, *bloks):
@@ -194,19 +167,12 @@ class Blok:
         :param blok: blok name
         :rtype: boolean
         """
-        Association = cls.registry.System.Blok.Association
-        total_association_count = Association.query().filter(
-            Association.blok == blok,
-            Association.mode == 'conditional').count()
-        query = Association.query().filter(
-            Association.blok == blok,
-            Association.mode == 'conditional')
-        query = query.join(cls, cls.name == Association.linked_blok)
-        query = query.filter(
-            cls.state.in_(['installed', 'toinstall', 'toupdate']))
-        association_count = query.count()
-        if total_association_count:
-            if total_association_count == association_count:
+        conditional = BlokManager.bloks[blok].conditional
+        if conditional:
+            query = cls.query().filter(cls.name.in_(conditional))
+            query = query.filter(
+                cls.state.in_(['installed', 'toinstall', 'toupdate']))
+            if len(conditional) == query.count():
                 return True
 
         return False
@@ -253,19 +219,3 @@ class Blok:
         bloks = query.order_by(cls.order).all()
         if bloks:
             bloks.load()
-
-
-@register(System.Blok)
-class Association:
-
-    MODES = {
-        'required': 'Required',
-        'optional': 'Optional',
-        'conditional': 'Conditional',
-    }
-
-    blok = String(foreign_key=System.Blok.use('name'),
-                  nullable=False, primary_key=True)
-    linked_blok = String(foreign_key=System.Blok.use('name'),
-                         nullable=False, primary_key=True)
-    mode = Selection(selections=MODES, nullable=False, primary_key=True)